feat: Add comprehensive social network features
- Comments system: Model, Repository, and PostDetailScreen
- Follow/Unfollow system: Repository, UI, Following/Followers screens
- Push Notifications: FCM integration with local notifications
- Home Screen: Central hub with personalized content
- Gamification: Points, levels, badges system
- Leaderboards: Multiple types with filtering
- Game Chat: Real-time chat for games
- Private Messages: Full messaging system

Security:
- Removed exposed Google Maps API keys
- Replaced with placeholders

Cleanup:
- Removed old backup files (.old)
- Updated .gitignore

+++ lib/utils/geohash_utils.dart
@@ -2,7 +2,7 @@
 /// Based on: https://en.wikipedia.org/wiki/Geohash
 class GeohashUtils {
   static const String _base32 = '0123456789bcdefghjkmnpqrstuvwxyz';
-  static const List<double> _bits = [16, 8, 4, 2, 1];
+  static const List<int> _bits = [16, 8, 4, 2, 1];
 
   /// Encode latitude and longitude to geohash
   static String encode(double latitude, double longitude, {int precision = 8}) {
